+++ package/example/lib/main.dart
@@ -37,6 +37,8 @@ class MyApp extends StatelessWidget {
       home: const HomePage(),
       routes: {
         "basic": (_) => const BasicExamplePage(),
+        "basic_pip": (_) => const BasicPipExamplePage(),
+        "basic_lock": (_) => const BasicLockControlsExamplePage(),
         "basic_with_looping": (_) => const BasicExampleWithLoopingPage(),
         "only_gestures": (_) => const OnlyGesturesExamplePage(),
         "secondary_controls": (_) => const SecondaryExamplePage(),
@@ -105,6 +107,19 @@ class _HomePageState extends State<HomePage> {
                       routeName: 'basic',
                       description:
                           'An example of how to load a video from a network source.',
+                    ),
+                    buildButton(
+                      context,
+                      text: 'Basic Pip Network example',
+                      routeName: 'basic_pip',
+                      description:
+                          'An example of how to load a video from a network source.',
+                    ),
+                    buildButton(
+                      context,
+                      text: 'Basic lock Network example',
+                      routeName: 'basic_lock',
+                      description: 'An example of which you can lock controls.',
                     ),
                     buildButton(
                       context,