+++ package/example/lib/pages/change_quality_example_page.dart
@@ -89,13 +89,13 @@ class _ChangeQualityExamplePageState extends State<ChangeQualityExamplePage> {
               onPressed: () {
                 _quality.value = quality; // change the video quality
                 _setDataSource(); // update the datasource
-                Navigator.pop(_);
+                Navigator.maybePop(_);
               },
             );
           },
         ),
         cancelButton: CupertinoActionSheetAction(
-          onPressed: () => Navigator.pop(_),
+          onPressed: () => Navigator.maybePop(_),
           isDestructiveAction: true,
           child: const Text("Cancel"),
         ),