+++ package/example/lib/pages/custom_controls.dart
@@ -13,8 +13,9 @@ class CustomControlsExamplePage extends StatefulWidget {
 
 class _CustomControlsExamplePageState extends State<CustomControlsExamplePage> {
   final _meeduPlayerController = MeeduPlayerController(
-      controlsStyle: ControlsStyle.custom,
-      enabledOverlays: const EnabledOverlays(volume: false));
+    controlsStyle: ControlsStyle.custom,
+    enabledOverlays: const EnabledOverlays(volume: false),
+  );
 
   StreamSubscription? _playerEventSubs;