+++ package/lib/src/controller.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:math';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_meedu/meedu.dart';
@@ -10,11 +11,18 @@ import 'package:flutter_meedu_media_kit/meedu_player.dart';
 import 'package:volume_controller/volume_controller.dart';
 import 'package:wakelock/wakelock.dart';
 import 'package:universal_platform/universal_platform.dart';
+import 'package:window_manager/window_manager.dart';
 
 /// An enumeration of the different styles that can be applied to controls, such
 /// as buttons and icons and layouts.
 enum ControlsStyle {
   primary,
+
+  /// When a video is inserted into a scrollable list, the scroll functionality
+  /// is disabled due to the drag event.
+  /// To address this, I added a style to the default controller to allow scrolling.
+  /// prevent drag event for scrollable list
+  primaryList,
   secondary,
 
   /// The custom style is used to apply a custom style which you can provide in MeeduPlayerController.
@@ -85,7 +93,7 @@ class MeeduPlayerController {
   Rx<bool> bufferingVideoDuration = false.obs;
 
   Rx<bool> videoFitChanged = false.obs;
-  final Rx<BoxFit> _videoFit = Rx(BoxFit.fill);
+  final Rx<BoxFit> _videoFit;
   //Rx<double> scale = 1.0.obs;
   Rx<bool> rewindIcons = false.obs;
   Rx<bool> forwardIcons = false.obs;
@@ -207,6 +215,9 @@ class MeeduPlayerController {
   /// for defining that video player needs mobile controls (even if its running on a web on a mobile device)
   bool mobileControls = false;
 
+  /// for defining that video player locked controls
+  final Rx<bool> _lockedControls = false.obs;
+
   /// controls if widgets inside videoplayer should get focus or not
   final bool excludeFocus;
 
@@ -230,6 +241,9 @@ class MeeduPlayerController {
 
   SharedPreferences? prefs;
 
+  DesktopPipBk? _desktopPipBk;
+  Size? _screenSizeBk;
+
   // returns the os version
   Future<double> get osVersion async {
     return _pipManager.osVersion;
@@ -240,6 +254,9 @@ class MeeduPlayerController {
 
   /// return fit of the Video,By default it is set to [BoxFit.contain]
   Rx<BoxFit> get videoFit => _videoFit;
+
+  /// returns true if the pip mode can used on the current device, the initial value will be false after check if pip is available
+  Rx<bool> get lockedControls => _lockedControls;
 
   /// A utility class that helps make the UI responsive by defining the size of
   /// icons, buttons, and text relative to the screen size.
@@ -252,12 +269,27 @@ class MeeduPlayerController {
   ///  this.buttonsSizeRelativeToScreen = 8,
   ///  this.maxButtonsSize = 40,
   ///});
-
   Responsive responsive = Responsive();
 
+  /// Defines the animation durations for various animations in a video.
+  ///
+  /// This class allows you to customize the duration of animations within the video player,
+  /// such as fade-in and fade-out durations, overlay show/hide animations, and more.
+  /// By modifying these durations, you can adjust the visual appearance and behavior
+  /// of the video player's animations according to your preferences.
   final Durations durations;
 
+  /// Controls the visibility of player overlays.
+  ///
+  /// Use this class to enable or disable the visibility of player overlays, such as
+  /// volume and brightness.
   final EnabledOverlays enabledOverlays;
+
+  /// Provides custom callback functions for specific player interactions.
+  ///
+  /// This class allows you to specify custom callback functions for player
+  /// interactions, such as long press events
+  final CustomCallbacks customCallbacks;
 
   /// creates an instance of [MeeduPlayerController]
   ///
@@ -272,7 +304,6 @@ class MeeduPlayerController {
     this.controlsEnabled = true,
     this.manageWakeLock = true,
     this.manageBrightness = true,
-    //TODOOOOOOO:
     this.showLogs = true,
     this.excludeFocus = true,
     String? errorText,
@@ -292,10 +323,12 @@ class MeeduPlayerController {
     this.enabledButtons = const EnabledButtons(),
     this.enabledControls = const EnabledControls(),
     this.enabledOverlays = const EnabledOverlays(),
+    this.customCallbacks = const CustomCallbacks(),
     Responsive? responsive,
     this.durations = const Durations(),
     this.onVideoPlayerClosed,
-  }) {
+    BoxFit? initialFit,
+  }) : _videoFit = Rx(initialFit ?? BoxFit.fill) {
     if (responsive != null) {
       this.responsive = responsive;
     }
@@ -303,7 +336,10 @@ class MeeduPlayerController {
     if (!manageBrightness) {
       enabledControls = enabledControls.copyWith(brightnessSwipes: false);
     }
-    getUserPreferenceForFit();
+
+    if (initialFit == null) {
+      getUserPreferenceForFit();
+    }
 
     _errorText = errorText;
     tag = DateTime.now().microsecondsSinceEpoch.toString();
@@ -344,15 +380,18 @@ class MeeduPlayerController {
       },
     );
 
-    if (pipEnabled && UniversalPlatform.isAndroid) {
-      // get the OS version and check if pip is available
-      _pipManager.checkPipAvailable().then(
-            (value) => _pipAvailable.value = value,
-          );
-      // listen the pip mode changes
-      _pipModeWorker = _pipManager.isInPipMode.ever(_onPipModeChanged);
-    } else {
-      _pipAvailable.value = false;
+    _pipAvailable.value = false;
+    if (pipEnabled) {
+      if (UniversalPlatform.isAndroid) {
+        // get the OS version and check if pip is available
+        _pipManager.checkPipAvailable().then(
+              (value) => _pipAvailable.value = value,
+            );
+        // listen the pip mode changes
+        _pipModeWorker = _pipManager.isInPipMode.ever(_onPipModeChanged);
+      } else if (UniversalPlatform.isDesktop) {
+        _pipAvailable.value = true;
+      }
     }
   }
 
@@ -754,6 +793,10 @@ class MeeduPlayerController {
 
   /// show or hide the player controls
   set controls(bool visible) {
+    // if (!UniversalPlatform.isDesktopOrWeb && visible && lockedControls.value) {
+    //   return;
+    // }
+
     // customDebugPrint("controls called with value $visible");
     if (fullscreen.value) {
       //customDebugPrint("Closed");
@@ -764,6 +807,12 @@ class MeeduPlayerController {
     _timer?.cancel();
     if (visible) {
       _hideTaskControls();
+    }
+  }
+
+  void toggleLockScreenMobile() {
+    if (!UniversalPlatform.isDesktopOrWeb) {
+      _lockedControls.value = !_lockedControls.value;
     }
   }
 
@@ -790,12 +839,23 @@ class MeeduPlayerController {
         screenManager.setWebFullScreen(true, this);
       } else {
         if (desktopOrWeb) {
+          if (!isInPipMode.value) {
+            _screenSizeBk = await windowManager.getSize();
+          }
           screenManager.setWindowsFullScreen(true, this);
         } else {
           screenManager.setFullScreenOverlaysAndOrientations();
         }
       }
     }
+    setVideoAsAppFullScreen(context,
+        applyOverlaysAndOrientations: applyOverlaysAndOrientations,
+        disposePlayer: disposePlayer);
+  }
+
+  Future<void> setVideoAsAppFullScreen(BuildContext context,
+      {bool applyOverlaysAndOrientations = true,
+      bool disposePlayer = false}) async {
     _fullscreen.value = true;
 
     final route = PageRouteBuilder(
@@ -809,7 +869,7 @@ class MeeduPlayerController {
       },
     );
 
-    await Navigator.push(context, route);
+    await Navigator.of(context).push(route);
   }
 
   /// launch a video using the fullscreen apge
@@ -997,7 +1057,7 @@ class MeeduPlayerController {
   void _onPipModeChanged(bool isInPipMode) {
     // if the pip mode was closed and before enter to pip mode the player was not in fullscreen
     if (!isInPipMode && _pipContextToFullscreen != null) {
-      Navigator.pop(_pipContextToFullscreen!); // close the fullscreen
+      Navigator.maybePop(_pipContextToFullscreen!); // close the fullscreen
       _pipContextToFullscreen = null;
     }
   }*/
@@ -1049,19 +1109,70 @@ class MeeduPlayerController {
     });
   }
 
+  double getAspectRatio() {
+    if (_videoPlayerController == null) {
+      return 16 / 9;
+    }
+
+    return _videoPlayerController!.value.size.width /
+        _videoPlayerController!.value.size.height;
+  }
+
   /// enter to picture in picture mode only Android
   ///
   /// only available since Android 7
   Future<void> enterPip(BuildContext context) async {
     if (pipAvailable.value && pipEnabled) {
-      controls = false; // hide the controls
-      if (!fullscreen.value) {
-        // if the player is not in the fullscreen mode
-        _pipContextToFullscreen = context;
-        goToFullscreen(context, applyOverlaysAndOrientations: false);
+      if (UniversalPlatform.isAndroid) {
+        await _enterPipAndroid(context);
+      } else if (UniversalPlatform.isDesktop) {
+        await _enterPipDesktop(context);
       }
-      await _pipManager.enterPip();
     }
+  }
+
+  Future<void> _enterPipAndroid(BuildContext context) async {
+    controls = false; // hide the controls
+    if (!fullscreen.value) {
+      // if the player is not in the fullscreen mode
+      _pipContextToFullscreen = context;
+      goToFullscreen(context, applyOverlaysAndOrientations: false);
+    }
+    await _pipManager.enterPip();
+  }
+
+  Future<void> _enterPipDesktop(BuildContext context) async {
+    if (_videoPlayerController == null) return;
+    if (!fullscreen.value) {
+      setVideoAsAppFullScreen(context);
+    }
+    double minH = max(MediaQuery.of(context).size.height * 0.15, 200);
+    double defaultH = max(MediaQuery.of(context).size.height * 0.30, 400);
+
+    double aspectRatio = getAspectRatio();
+    _desktopPipBk = DesktopPipBk(
+      isFullScreen: await windowManager.isFullScreen(),
+      size: await windowManager.isFullScreen()
+          ? (_screenSizeBk ??
+              Size(
+                MediaQuery.of(context).size.width / 2,
+                MediaQuery.of(context).size.height / 2,
+              ))
+          : await windowManager.getSize(),
+    );
+
+    await onFullscreenClose();
+    // ignore: use_build_context_synchronously
+
+    await windowManager.setTitleBarStyle(TitleBarStyle.hidden);
+    await windowManager.center(animate: true);
+    await windowManager.setAlwaysOnTop(true);
+    await windowManager.setMinimumSize(Size(minH * (aspectRatio), minH));
+    await windowManager.setSize(Size(defaultH * (aspectRatio), defaultH));
+    // await windowManager.setAsFrameless();
+    await windowManager.setAspectRatio(aspectRatio);
+    // windowManager.setSkipTaskbar(true);
+    _pipManager.isInPipMode.value = true;
   }
 
   /// listener for pip changes
@@ -1071,6 +1182,38 @@ class MeeduPlayerController {
       Navigator.pop(_pipContextToFullscreen!); // close the fullscreen
       _pipContextToFullscreen = null;
     }
+  }
+
+  void closePip(BuildContext context) {
+    if (_pipManager.isInPipMode.value == true) {
+      if (UniversalPlatform.isDesktop) {
+        if (!_desktopPipBk!.isFullScreen) {
+          // ignore: use_build_context_synchronously
+          Navigator.of(context).pop();
+        }
+
+        _closePipDesktop(context);
+      }
+    }
+  }
+
+  Future<void> _closePipDesktop(BuildContext context) async {
+    double defaultSizeHeight =
+        max(MediaQuery.of(context).size.height * 0.30, 300);
+    double defaultSizeWidth =
+        max(MediaQuery.of(context).size.width * 0.30, 500);
+
+    await windowManager.setTitleBarStyle(TitleBarStyle.normal);
+    await windowManager.setAlwaysOnTop(false);
+    await windowManager.setAspectRatio(0);
+    // // windowManager.setSkipTaskbar(false);
+    await windowManager.setSize(_desktopPipBk!.size);
+    await windowManager
+        .setMinimumSize(Size(defaultSizeWidth, defaultSizeHeight));
+    if (_desktopPipBk!.isFullScreen) {
+      screenManager.setWindowsFullScreen(true, this);
+    }
+    _pipManager.isInPipMode.value = false;
   }
 
   static MeeduPlayerController of(BuildContext context) {